+++ NEWS.md
@@ -3,6 +3,7 @@
 - switch to Github Actions for CI
 - units package now supported (#51)
 - mutate: fix formatting issues for NAs (#48)
+- use rlang::inform instead of message (#41)
 
 # 1.0.2